refactor orchestration to be closer to other API, add private repository creds to create/delete
