+++ ecs/ecs_test.go
@@ -5,7 +5,22 @@ import (
 	"testing"
 
 	"github.com/YaleSpinup/ecs-api/common"
+	"github.com/aws/aws-sdk-go/service/ecs/ecsiface"
 )
+
+// mockECSClient is a fake ecs client
+type mockECSClient struct {
+	ecsiface.ECSAPI
+	t   *testing.T
+	err error
+}
+
+func newmockECSClient(t *testing.T, err error) ecsiface.ECSAPI {
+	return &mockECSClient{
+		t:   t,
+		err: err,
+	}
+}
 
 func TestNewSession(t *testing.T) {
 	e := NewSession(common.Account{})
