+++ orchestration/repositorycredentials_test.go
@@ -1,0 +1,9 @@
+package orchestration
+
+import (
+	"testing"
+)
+
+func TestProcessRepositoryCredentials(t *testing.T) {
+
+}
